requirements.txt
numpy
opencv-python-headless
matplotlib
Pillow
scipy
imutils
